requirements.txt
acres==0.5.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.4
attrs==25.1.0
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
cloudscraper==1.2.71
comm==0.2.2
configobj==5.0.9
configparser==7.2.0
cryptography==45.0.4
Cython==3.0.11
dataclasses-json==0.6.7
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
dill==0.4.0
distlib==0.3.9
distro==1.9.0
Django==5.1.5
django-phonenumber-field==8.1.0
django-widget-tweaks==1.5.0
etelemetry==0.3.1
executing==2.2.0
exejs==0.0.4
fastjsonschema==2.21.1
filelock==3.13.1
fitz==0.0.1.dev2
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.2.0
git-filter-repo==2.34.0
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
h11==0.14.0
hf-xet==1.1.5
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
isoduration==20.11.0
jedi==0.19.2
jh2==5.0.9
Jinja2==3.1.3
jiter==0.8.2
joblib==1.5.1
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
langchain==0.3.17
langchain-anthropic==0.3.5
langchain-community==0.3.16
langchain-core==0.3.33
langchain-openai==0.3.3
langchain-text-splitters==0.3.5
langdetect==1.0.9
langsmith==0.3.5
looseversion==1.3.0
lxml==5.4.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib-inline==0.1.7
mistune==3.1.1
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.18
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nibabel==5.3.2
nipype==1.10.0
niquests==3.14.1
nltk==3.9.1
notebook==7.3.2
notebook_shim==0.2.4
numpy==2.2.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openai==1.61.0
openfoodfacts==2.5.0
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.3.0
pandocfilters==1.5.1
parso==0.8.4
pathlib==1.0.1
pathos==0.3.4
pexpect==4.9.0
phonenumberslite==9.0.8
pillow==11.1.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.6
pox==0.3.6
ppft==1.7.7
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
prov==2.1.1
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.29
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydot==4.0.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==26.2.1
qh3==1.5.3
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
Send2Trash==1.8.3
sentence-transformers==4.1.0
setuptools==70.0.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
sqlparse==0.5.3
stack-data==0.6.3
svgwrite==1.4.3
sympy==1.14.0
tenacity==9.0.0
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.8.0
tinycss2==1.4.0
tldextract==5.3.0
tokenizers==0.21.2
torch==2.7.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
transformers==4.52.4
translations==0.1.3
translators==6.0.1
Tree==0.2.4
triton==3.3.1
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.3.0
urllib3-future==2.13.900
virtualenv==20.29.1
wassima==1.2.2
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.45.1
whitenoise==6.8.2
yarl==1.18.3
zstandard==0.23.0
